+++ src/ctl/src/cmd_impl/await_tree.rs
@@ -12,12 +12,96 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod analyze;
-mod dump;
-mod transcribe;
-mod tree;
-mod utils;
+use risingwave_common::util::StackTraceResponseExt;
+use risingwave_common::util::addr::HostAddr;
+use risingwave_pb::common::WorkerType;
+use risingwave_pb::monitor_service::{StackTraceRequest, StackTraceResponse};
+use risingwave_rpc_client::{CompactorClient, ComputeClientPool};
+use rw_diagnose_tools::await_tree::TreeView;
 
-pub use analyze::*;
-pub use dump::*;
-pub use transcribe::*;
+use crate::CtlContext;
+
+pub async fn dump(context: &CtlContext, actor_traces_format: Option<String>) -> anyhow::Result<()> {
+    let mut all = StackTraceResponse::default();
+
+    let meta_client = context.meta_client().await?;
+
+    let compute_nodes = meta_client
+        .list_worker_nodes(Some(WorkerType::ComputeNode))
+        .await?;
+    let clients = ComputeClientPool::adhoc();
+
+    let req = StackTraceRequest {
+        actor_traces_format,
+    };
+    // FIXME: the compute node may not be accessible directly from risectl, we may let the meta
+    // service collect the reports from all compute nodes in the future.
+    for cn in compute_nodes {
+        let client = clients.get(&cn).await?;
+        let response = client.stack_trace(req.clone()).await?;
+        all.merge_other(response);
+    }
+
+    let compactor_nodes = meta_client
+        .list_worker_nodes(Some(WorkerType::Compactor))
+        .await?;
+
+    for compactor in compactor_nodes {
+        let addr: HostAddr = compactor.get_host().unwrap().into();
+        let client = CompactorClient::new(addr).await?;
+        let response = client.stack_trace(req.clone()).await?;
+        all.merge_other(response);
+    }
+
+    if all.actor_traces.is_empty()
+        && all.rpc_traces.is_empty()
+        && all.compaction_task_traces.is_empty()
+        && all.inflight_barrier_traces.is_empty()
+    {
+        println!("No traces found. No actors are running, or `--async-stack-trace` not set?");
+    }
+    println!("{}", all.output());
+
+    Ok(())
+}
+
+pub async fn bottleneck_detect(context: &CtlContext, path: Option<String>) -> anyhow::Result<()> {
+    if let Some(path) = path {
+        rw_diagnose_tools::await_tree::bottleneck_detect_from_file(path)
+    } else {
+        bottleneck_detect_real_time(context).await
+    }
+}
+
+async fn bottleneck_detect_real_time(context: &CtlContext) -> anyhow::Result<()> {
+    let meta_client = context.meta_client().await?;
+
+    let compute_nodes = meta_client
+        .list_worker_nodes(Some(WorkerType::ComputeNode))
+        .await?;
+    let clients = ComputeClientPool::adhoc();
+
+    // request for json actor traces
+    let req = StackTraceRequest::default();
+
+    let mut bottleneck_actors_found = false;
+    for cn in compute_nodes {
+        let client = clients.get(&cn).await?;
+        let response = client.stack_trace(req.clone()).await?;
+        for (actor_id, trace) in response.actor_traces {
+            let tree: TreeView = serde_json::from_str(&trace)
+                .map_err(|e| anyhow::anyhow!("Failed to parse actor trace JSON: {}", e))?;
+            if tree.is_bottleneck() {
+                bottleneck_actors_found = true;
+                println!(">> Actor {}", actor_id);
+                println!("{}", tree);
+            }
+        }
+    }
+
+    if !bottleneck_actors_found {
+        println!("No bottleneck actors detected.");
+    }
+
+    Ok(())
+}
